Added interactive tests, refined logic
