--- .percyrc
@@ -1,5 +1,0 @@
-version: 1
-snapshot:
-  widths: [1280]
-  min-height: 1024 # px
-  enable-javascript: true
