+++ package.json
@@ -17,7 +17,7 @@
     "test:local": "KARMA_SINGLE_RUN=false karma start ./src/tests/config/karma.conf.js",
     "test": "karma start ./src/tests/config/karma.conf.nomodule.js && karma start ./src/tests/config/karma.conf.js && codecov",
     "test:browserstack": " TEST_ON_BROWSERSTACK=true karma start ./src/tests/config/karma.conf.nomodule.js && TEST_ON_BROWSERSTACK=true karma start ./src/tests/config/karma.conf.js",
-    "test-visual": "yarn build && npx percy exec -- babel-node src/tests/visual/percy.snapshot.js",
+    "test-visual": "yarn build && npx percy exec -- babel-node src/tests/visual/percy.snapshots.js",
     "tidy-clean": "rm -rf build css favicons fonts img components page-templates templates scripts coverage scss",
     "check-unused": "npx npm-check-unused",
     "dedupe-deps": "npx yarn-deduplicate yarn.lock",
